Change array type

+++ evalAssignment1/update_locations.c
@@ -3,7 +3,7 @@
 // Create a list of 'size' floating point numbers in the range [-bound, bound]
 float * generate_random_list(int32_t size, int32_t bound) {
     float* array;
-    array = (float*)malloc(sizeof(int32_t) * size);
+    array = (float*)malloc(sizeof(float) * size);
 
 
     for (int i = 0; i < size; i++) {
